First example displaying the geometry in 3D
The geometry.C example script reads the acceptance from the Delphes
configuration and generates from there a 3D representation of the
detector in ROOT (using the geometry package).

This initial version makes few assumptions: the detector must be
symmetric and have a single calorimeter. The module names are fixed.
Therefore, the script works with CMS and ATLAS geometries but not with
the prototype FCC geometry.