Display of automatically selected elements working
The original features are back, but with a much more flexible code that
selects automatically the list of branches from the tcl file. The main
issue left is the GUI (handler not found), and then some more
cleaning. One idea is to further exploit templating to avoid ifs when
leading events.