Add Invert and Status options to PdgCodeFilter
It is sometimes useful to output the truth record of a single
(non-stable) particle species, for example when considering an
exotic heavy resonance. Presently the only way to do so is to
output the entire Delphes/allParticles array, which is inefficient.

This patch adds the option `InvertPdg` to PdgCodeFilter, which causes
the PdgCodes specified to be *added* to the output array, rather
than removed. It also adds the options `RequireStatus` and `Status`,
for additional filtering based on the MC record status. If
`RequireStatus` is `true`, than only particles matching `Status` are
added to the output array.

This patch should be backwards compatible with existing usages of
PdgCodeFilter.