Change candidate momentum to reflect the perigee momentum
In the particle propagator the candidate momentum is currently copied
from the truth particle momentum. This is correct for prompt
particles, but inconsistent with most (all?) collider experiments,
which use the perigee momentum. The effect is especially noticeable in
low-pt tracks coming from a highly displaced vertex.

This commit fixes this by adjusting the candidate px and py, and by
extension the candidate dxy. The discrepancy should have a minimal
effect on most quantities, but could prevent problems which arise if
the user adds any vertex reconstruction algorithms to the standard
Delphes modules.